api/requirements.txt
flask
flask_sqlalchemy
flask-login
flask-wtf
flask_jwt_extended
wtforms
sqlalchemy_utils
requests
psycopg2-binary
python3-discogs-client